+++ prisma/migrations/20260827200430_add_osm_way_id_to_road_segment/migration.sql
@@ -1,0 +1,11 @@
+/*
+  Warnings:
+
+  - A unique constraint covering the columns `[osmWayId]` on the table `RoadSegment` will be added. If there are existing duplicate values, this will fail.
+
+*/
+-- AlterTable
+ALTER TABLE "RoadSegment" ADD COLUMN     "osmWayId" BIGINT;
+
+-- CreateIndex
+CREATE UNIQUE INDEX "RoadSegment_osmWayId_key" ON "RoadSegment"("osmWayId");